requirements.txt
anthropic>=0.40,<1.0
beautifulsoup4>=4.12
google-ads>=24.0
playwright>=1.45
pillow>=10.0
python-dotenv>=1.0
PyYAML>=6.0
requests>=2.32
flask>=3.0
resend>=2.0
pytest>=8.0
